Refactor opaque render into specialized passes
