+++ viewer3d/viewer3dcontroller.cpp
@@ -55,6 +55,9 @@
 #include "projectutils.h"
 #include "render/scenerenderer.h"
 #include "render/render_pipeline.h"
+#include "render/opaque_fixture_pass.h"
+#include "render/opaque_object_pass.h"
+#include "render/opaque_truss_pass.h"
 #include "culling/visibilitysystem.h"
 #include "labels/label_render_system.h"
 #include "picking/selectionsystem.h"
@@ -1296,8 +1299,6 @@ void Viewer3DController::RenderOpaqueFrame(const RenderFrameContext &context,
   const bool wireframe = context.wireframe;
   const Viewer2DRenderMode mode = context.mode;
   const Viewer2DView view = context.view;
-  const bool skipCapture = context.skipCapture;
-  const bool is2DViewer = context.is2DViewer;
 
   if (context.useLighting)
     SetupBasicLighting();
@@ -1310,7 +1311,6 @@ void Viewer3DController::RenderOpaqueFrame(const RenderFrameContext &context,
 
   auto getTypeColor = [&](const std::string &key, const std::string &hex) {
     std::array<float, 3> c;
-    // Always honor user-specified colors, updating the cache if needed
     if (!hex.empty() && HexToRGB(hex, c[0], c[1], c[2])) {
       m_typeColors[key] = c;
       return c;
@@ -1344,629 +1344,12 @@ void Viewer3DController::RenderOpaqueFrame(const RenderFrameContext &context,
     }
   };
 
-  const auto &sceneObjects = SceneDataManager::Instance().GetSceneObjects();
-  const auto &trusses = SceneDataManager::Instance().GetTrusses();
-  const auto &fixtures = SceneDataManager::Instance().GetFixtures();
-
-  // Scene objects first
-  glShadeModel(GL_FLAT);
-  for (const auto &uuid : visibleSet.objectUuids) {
-    auto sceneIt = sceneObjects.find(uuid);
-    if (sceneIt == sceneObjects.end())
-      continue;
-    const auto &m = sceneIt->second;
-    glPushMatrix();
-
-    std::string objectCaptureKey;
-    if (m_captureCanvas && !skipCapture) {
-      objectCaptureKey = m.modelFile.empty() ? m.name : m.modelFile;
-      if (objectCaptureKey.empty())
-        objectCaptureKey = "scene_object";
-      m_captureCanvas->SetSourceKey(objectCaptureKey);
-    }
-
-    bool highlight = (!m_highlightUuid.empty() && uuid == m_highlightUuid);
-    bool selected = (m_selectedUuids.find(uuid) != m_selectedUuids.end());
-
-    float matrix[16];
-    MatrixToArray(m.transform, matrix);
-    ApplyTransform(matrix, true);
-
-    float cx = 0.0f, cy = 0.0f, cz = 0.0f;
-    auto obit = m_objectBounds.find(uuid);
-    if (obit != m_objectBounds.end()) {
-      cx = (obit->second.min[0] + obit->second.max[0]) * 0.5f;
-      cy = (obit->second.min[1] + obit->second.max[1]) * 0.5f;
-      cz = (obit->second.min[2] + obit->second.max[2]) * 0.5f;
-      cx -= m.transform.o[0] * RENDER_SCALE;
-      cy -= m.transform.o[1] * RENDER_SCALE;
-      cz -= m.transform.o[2] * RENDER_SCALE;
-    }
-
-    float r = 1.0f, g = 1.0f, b = 1.0f;
-    if (wireframe && mode == Viewer2DRenderMode::ByLayer) {
-      auto c = getLayerColor(m.layer);
-      r = c[0];
-      g = c[1];
-      b = c[2];
-    }
-
-    Matrix captureTransform = m.transform;
-    captureTransform.o[0] *= RENDER_SCALE;
-    captureTransform.o[1] *= RENDER_SCALE;
-    captureTransform.o[2] *= RENDER_SCALE;
-    auto applyCapture = [captureTransform](const std::array<float, 3> &p) {
-      return TransformPoint(captureTransform, p);
-    };
-
-    struct SceneObjectMeshPart {
-      const Mesh *mesh = nullptr;
-      Matrix localTransform = MatrixUtils::Identity();
-      std::string modelKey;
-    };
-    std::vector<SceneObjectMeshPart> objectMeshParts;
-    if (!m.geometries.empty()) {
-      for (const auto &geo : m.geometries) {
-        std::string objectPath;
-        auto pathIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(geo.modelFile));
-        if (pathIt != m_resourceSyncState.resolvedModelRefs.end() && pathIt->second.attempted)
-          objectPath = pathIt->second.resolvedPath;
-        if (objectPath.empty())
-          continue;
-        auto it = m_resourceSyncState.loadedMeshes.find(objectPath);
-        if (it == m_resourceSyncState.loadedMeshes.end())
-          continue;
-
-        SceneObjectMeshPart part;
-        part.mesh = &it->second;
-        part.localTransform = geo.localTransform;
-        part.modelKey = NormalizeModelKey(objectPath);
-        objectMeshParts.push_back(std::move(part));
-      }
-    } else if (!m.modelFile.empty()) {
-      std::string objectPath;
-      auto pathIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(m.modelFile));
-      if (pathIt != m_resourceSyncState.resolvedModelRefs.end() && pathIt->second.attempted)
-        objectPath = pathIt->second.resolvedPath;
-      if (!objectPath.empty()) {
-        auto it = m_resourceSyncState.loadedMeshes.find(objectPath);
-        if (it != m_resourceSyncState.loadedMeshes.end()) {
-          SceneObjectMeshPart part;
-          part.mesh = &it->second;
-          part.modelKey = NormalizeModelKey(objectPath);
-          objectMeshParts.push_back(std::move(part));
-        }
-      }
-    }
-
-    auto drawSceneObjectGeometry =
-        [&](const std::function<std::array<float, 3>(
-                const std::array<float, 3> &)> &captureTransform,
-            bool isHighlighted, bool isSelected) {
-          if (!objectMeshParts.empty()) {
-            for (const auto &part : objectMeshParts) {
-              Matrix worldMatrix = MatrixUtils::Multiply(m.transform, part.localTransform);
-              float partMatrix[16];
-              MatrixToArray(worldMatrix, partMatrix);
-
-              Matrix partCaptureMatrix = worldMatrix;
-              partCaptureMatrix.o[0] *= RENDER_SCALE;
-              partCaptureMatrix.o[1] *= RENDER_SCALE;
-              partCaptureMatrix.o[2] *= RENDER_SCALE;
-              auto partCapture = [partCaptureMatrix](const std::array<float, 3> &p) {
-                return TransformPoint(partCaptureMatrix, p);
-              };
-
-              float localMatrix[16];
-              MatrixToArray(part.localTransform, localMatrix);
-              glPushMatrix();
-              ApplyTransform(localMatrix, false);
-              auto partCaptureTransform = captureTransform;
-              if (captureTransform)
-                partCaptureTransform = partCapture;
-
-              DrawMeshWithOutline(*part.mesh, r, g, b, RENDER_SCALE,
-                                  isHighlighted, isSelected, cx, cy, cz,
-                                  wireframe, mode, partCaptureTransform, false,
-                                  partMatrix);
-              glPopMatrix();
-            }
-          } else {
-            DrawCubeWithOutline(0.3f, r, g, b, isHighlighted, isSelected, cx,
-                                cy, cz, wireframe, mode, captureTransform);
-          }
-        };
-
-    bool suppressCapture = false;
-    const bool useSymbolInstancing =
-        (m_captureUseSymbols &&
-         (m_captureView == Viewer2DView::Bottom ||
-          m_captureView == Viewer2DView::Top ||
-          m_captureView == Viewer2DView::Front ||
-          m_captureView == Viewer2DView::Side) &&
-         !highlight && !selected);
-    bool placedInstance = false;
-    if (useSymbolInstancing && m_captureCanvas && !skipCapture) {
-      std::string modelKey;
-      if (!objectMeshParts.empty())
-        modelKey = objectMeshParts.front().modelKey;
-      else if (!m.modelFile.empty())
-        modelKey = NormalizeModelKey(m.modelFile);
-      if (modelKey.empty() && !m.name.empty())
-        modelKey = m.name;
-
-      if (!modelKey.empty()) {
-        SymbolKey symbolKey;
-        symbolKey.modelKey = "object:" + modelKey;
-        symbolKey.viewKind = resolveSymbolView(m_captureView);
-        symbolKey.styleVersion = 1;
-
-        const auto &symbol =
-            m_bottomSymbolCache.GetOrCreate(symbolKey, [&](const SymbolKey &,
-                                                           uint32_t symbolId) {
-              SymbolDefinition definition{};
-              definition.symbolId = symbolId;
-              auto localCanvas =
-                  CreateRecordingCanvas(definition.localCommands, false);
-              CanvasTransform transform{};
-              localCanvas->BeginFrame();
-              localCanvas->SetTransform(transform);
-
-              ICanvas2D *prevCanvas = m_captureCanvas;
-              Viewer2DView prevView = m_captureView;
-              bool prevCaptureOnly = m_captureOnly;
-              bool prevIncludeGrid = m_captureIncludeGrid;
-              m_captureCanvas = localCanvas.get();
-              m_captureView = prevView;
-              m_captureOnly = true;
-              m_captureIncludeGrid = false;
-
-              m_captureCanvas->SetSourceKey(
-                  objectCaptureKey.empty() ? "scene_object" : objectCaptureKey);
-              drawSceneObjectGeometry({}, false, false);
-              localCanvas->EndFrame();
-              definition.bounds =
-                  ComputeSymbolBounds(definition.localCommands);
-
-              m_captureCanvas = prevCanvas;
-              m_captureView = prevView;
-              m_captureOnly = prevCaptureOnly;
-              m_captureIncludeGrid = prevIncludeGrid;
-              return definition;
-            });
-
-        Transform2D instanceTransform =
-            BuildInstanceTransform2D(captureTransform, m_captureView);
-        m_captureCanvas->PlaceSymbolInstance(symbol.symbolId, instanceTransform);
-        placedInstance = true;
-      }
-    }
-    suppressCapture = placedInstance;
-
-    if (suppressCapture) {
-      ICanvas2D *prevCanvas = m_captureCanvas;
-      bool prevCaptureOnly = m_captureOnly;
-      m_captureCanvas = nullptr;
-      m_captureOnly = false;
-      drawSceneObjectGeometry(applyCapture, highlight, selected);
-      m_captureCanvas = prevCanvas;
-      m_captureOnly = prevCaptureOnly;
-    } else {
-      drawSceneObjectGeometry(applyCapture, highlight, selected);
-    }
-
-    glPopMatrix();
-  }
-
-  // Trusses next
-  glShadeModel(GL_SMOOTH); // keep smooth shading for trusses
-  for (const auto &uuid : visibleSet.trussUuids) {
-    auto trussIt = trusses.find(uuid);
-    if (trussIt == trusses.end())
-      continue;
-    const auto &t = trussIt->second;
-    glPushMatrix();
-
-    std::string trussCaptureKey;
-    if (m_captureCanvas && !skipCapture) {
-      trussCaptureKey = t.model.empty() ? t.name : t.model;
-      if (trussCaptureKey.empty())
-        trussCaptureKey = "truss";
-      m_captureCanvas->SetSourceKey(trussCaptureKey);
-    }
-
-    bool highlight = (!m_highlightUuid.empty() && uuid == m_highlightUuid);
-    bool selected = (m_selectedUuids.find(uuid) != m_selectedUuids.end());
-
-    float matrix[16];
-    MatrixToArray(t.transform, matrix);
-    ApplyTransform(matrix, true);
-
-    float cx = 0.0f, cy = 0.0f, cz = 0.0f;
-    auto tbit = m_trussBounds.find(uuid);
-    if (tbit != m_trussBounds.end()) {
-      cx = (tbit->second.min[0] + tbit->second.max[0]) * 0.5f;
-      cy = (tbit->second.min[1] + tbit->second.max[1]) * 0.5f;
-      cz = (tbit->second.min[2] + tbit->second.max[2]) * 0.5f;
-      cx -= t.transform.o[0] * RENDER_SCALE;
-      cy -= t.transform.o[1] * RENDER_SCALE;
-      cz -= t.transform.o[2] * RENDER_SCALE;
-    }
-
-    float r = 1.0f, g = 1.0f, b = 1.0f;
-    if (wireframe && mode == Viewer2DRenderMode::ByLayer) {
-      auto c = getLayerColor(t.layer);
-      r = c[0];
-      g = c[1];
-      b = c[2];
-    }
-
-    Matrix captureTransform = t.transform;
-    captureTransform.o[0] *= RENDER_SCALE;
-    captureTransform.o[1] *= RENDER_SCALE;
-    captureTransform.o[2] *= RENDER_SCALE;
-    auto applyCapture = [captureTransform](const std::array<float, 3> &p) {
-      return TransformPoint(captureTransform, p);
-    };
-
-    const Mesh *trussMesh = nullptr;
-    std::string trussPath;
-    if (!t.symbolFile.empty()) {
-      auto trussPathIt = m_resourceSyncState.resolvedModelRefs.find(ResolveCacheKey(t.symbolFile));
-      if (trussPathIt != m_resourceSyncState.resolvedModelRefs.end() && trussPathIt->second.attempted)
-        trussPath = trussPathIt->second.resolvedPath;
-      if (!trussPath.empty()) {
-        auto it = m_resourceSyncState.loadedMeshes.find(trussPath);
-        if (it != m_resourceSyncState.loadedMeshes.end())
-          trussMesh = &it->second;
-      }
-    }
-
-    float trussLen = t.lengthMm * RENDER_SCALE;
-    float trussWid = (t.widthMm > 0 ? t.widthMm : 400.0f) * RENDER_SCALE;
-    float trussHei = (t.heightMm > 0 ? t.heightMm : 400.0f) * RENDER_SCALE;
-    float trussWidthMm = (t.widthMm > 0 ? t.widthMm : 400.0f);
-    float trussHeightMm = (t.heightMm > 0 ? t.heightMm : 400.0f);
-
-    auto drawTrussGeometry =
-        [&](const std::function<std::array<float, 3>(
-                const std::array<float, 3> &)> &captureTransform,
-            bool isHighlighted, bool isSelected) {
-          if (trussMesh) {
-            DrawMeshWithOutline(*trussMesh, r, g, b, RENDER_SCALE,
-                                isHighlighted, isSelected, cx, cy, cz,
-                                wireframe, mode, captureTransform, false,
-                                matrix);
-          } else {
-            DrawWireframeBox(trussLen, trussHei, trussWid, isHighlighted,
-                             isSelected, wireframe, mode, captureTransform);
-          }
-        };
-
-    bool suppressCapture = false;
-    const bool useSymbolInstancing =
-        (m_captureUseSymbols &&
-         (m_captureView == Viewer2DView::Bottom ||
-          m_captureView == Viewer2DView::Top ||
-          m_captureView == Viewer2DView::Front ||
-          m_captureView == Viewer2DView::Side) &&
-         !highlight && !selected);
-    bool placedInstance = false;
-    if (useSymbolInstancing && m_captureCanvas && !skipCapture) {
-      std::string modelKey;
-      if (!trussPath.empty())
-        modelKey = NormalizeModelKey(trussPath);
-      else if (!t.symbolFile.empty())
-        modelKey = NormalizeModelKey(t.symbolFile);
-      if (modelKey.empty() && !trussMesh) {
-        std::ostringstream boxKey;
-        boxKey << "box:" << t.lengthMm << "x" << trussWidthMm << "x"
-               << trussHeightMm;
-        modelKey = boxKey.str();
-      }
-      if (modelKey.empty() && !t.model.empty())
-        modelKey = t.model;
-      if (modelKey.empty() && !t.name.empty())
-        modelKey = t.name;
-
-      if (!modelKey.empty()) {
-        SymbolKey symbolKey;
-        symbolKey.modelKey = "truss:" + modelKey;
-        symbolKey.viewKind = resolveSymbolView(m_captureView);
-        symbolKey.styleVersion = 1;
-
-        const auto &symbol =
-            m_bottomSymbolCache.GetOrCreate(symbolKey, [&](const SymbolKey &,
-                                                           uint32_t symbolId) {
-              SymbolDefinition definition{};
-              definition.symbolId = symbolId;
-              auto localCanvas =
-                  CreateRecordingCanvas(definition.localCommands, false);
-              CanvasTransform transform{};
-              localCanvas->BeginFrame();
-              localCanvas->SetTransform(transform);
-
-              ICanvas2D *prevCanvas = m_captureCanvas;
-              Viewer2DView prevView = m_captureView;
-              bool prevCaptureOnly = m_captureOnly;
-              bool prevIncludeGrid = m_captureIncludeGrid;
-              m_captureCanvas = localCanvas.get();
-              m_captureView = prevView;
-              m_captureOnly = true;
-              m_captureIncludeGrid = false;
-
-              m_captureCanvas->SetSourceKey(
-                  trussCaptureKey.empty() ? "truss" : trussCaptureKey);
-              drawTrussGeometry({}, false, false);
-              localCanvas->EndFrame();
-              definition.bounds =
-                  ComputeSymbolBounds(definition.localCommands);
-
-              m_captureCanvas = prevCanvas;
-              m_captureView = prevView;
-              m_captureOnly = prevCaptureOnly;
-              m_captureIncludeGrid = prevIncludeGrid;
-              return definition;
-            });
-
-        Transform2D instanceTransform =
-            BuildInstanceTransform2D(captureTransform, m_captureView);
-        m_captureCanvas->PlaceSymbolInstance(symbol.symbolId, instanceTransform);
-        placedInstance = true;
-      }
-    }
-    suppressCapture = placedInstance;
-
-    if (suppressCapture) {
-      ICanvas2D *prevCanvas = m_captureCanvas;
-      bool prevCaptureOnly = m_captureOnly;
-      m_captureCanvas = nullptr;
-      m_captureOnly = false;
-      drawTrussGeometry(applyCapture, highlight, selected);
-      m_captureCanvas = prevCanvas;
-      m_captureOnly = prevCaptureOnly;
-    } else {
-      drawTrussGeometry(applyCapture, highlight, selected);
-    }
-
-    glPopMatrix();
-  }
-
-  // Fixtures last
-  glShadeModel(GL_FLAT);
-  const bool forceFixturesOnTop = wireframe;
-  GLboolean depthEnabled = GL_FALSE;
-  if (forceFixturesOnTop) {
-    depthEnabled = glIsEnabled(GL_DEPTH_TEST);
-    if (depthEnabled)
-      glDisable(GL_DEPTH_TEST);
-  }
-  for (const auto &uuid : visibleSet.fixtureUuids) {
-    auto fixtureIt = fixtures.find(uuid);
-    if (fixtureIt == fixtures.end())
-      continue;
-    const auto &f = fixtureIt->second;
-    glPushMatrix();
-
-    std::string fixtureCaptureKey;
-    if (m_captureCanvas && !skipCapture) {
-      fixtureCaptureKey = !f.typeName.empty()
-                              ? f.typeName
-                              : (!f.gdtfSpec.empty() ? f.gdtfSpec : "unknown");
-      // Default to the whole fixture key; individual GDTF parts will override
-      // this key below so outlines remain visible in captured PDFs.
-      m_captureCanvas->SetSourceKey(fixtureCaptureKey);
-    }
-
-    bool highlight = (!m_highlightUuid.empty() && uuid == m_highlightUuid);
-    bool selected = (m_selectedUuids.find(uuid) != m_selectedUuids.end());
-
-    float matrix[16];
-    MatrixToArray(f.transform, matrix);
-    ApplyTransform(matrix, true);
-
-    float cx = 0.0f, cy = 0.0f, cz = 0.0f;
-    auto fbit = m_fixtureBounds.find(uuid);
-    if (fbit != m_fixtureBounds.end()) {
-      cx = (fbit->second.min[0] + fbit->second.max[0]) * 0.5f;
-      cy = (fbit->second.min[1] + fbit->second.max[1]) * 0.5f;
-      cz = (fbit->second.min[2] + fbit->second.max[2]) * 0.5f;
-      cx -= f.transform.o[0] * RENDER_SCALE;
-      cy -= f.transform.o[1] * RENDER_SCALE;
-      cz -= f.transform.o[2] * RENDER_SCALE;
-    }
-
-    float r = 1.0f, g = 1.0f, b = 1.0f;
-    if (wireframe) {
-      if (mode == Viewer2DRenderMode::ByFixtureType) {
-        auto c = getTypeColor(f.gdtfSpec, f.color);
-        r = c[0];
-        g = c[1];
-        b = c[2];
-      } else if (mode == Viewer2DRenderMode::ByLayer) {
-        auto c = getLayerColor(f.layer);
-        r = c[0];
-        g = c[1];
-        b = c[2];
-      }
-    }
-
-    Matrix fixtureTransform = f.transform;
-    fixtureTransform.o[0] *= RENDER_SCALE;
-    fixtureTransform.o[1] *= RENDER_SCALE;
-    fixtureTransform.o[2] *= RENDER_SCALE;
-
-    auto applyFixtureCapture = [fixtureTransform](
-                                   const std::array<float, 3> &p) {
-      return TransformPoint(fixtureTransform, p);
-    };
-
-    std::string gdtfPath;
-    auto gdtfPathIt = m_resourceSyncState.resolvedGdtfSpecs.find(ResolveCacheKey(f.gdtfSpec));
-    if (gdtfPathIt != m_resourceSyncState.resolvedGdtfSpecs.end() && gdtfPathIt->second.attempted)
-      gdtfPath = gdtfPathIt->second.resolvedPath;
-    auto itg = m_resourceSyncState.loadedGdtf.find(gdtfPath);
-
-    bool suppressCapture = false;
-    const bool useSymbolInstancing =
-        (m_captureUseSymbols &&
-         (m_captureView == Viewer2DView::Bottom ||
-          m_captureView == Viewer2DView::Top ||
-          m_captureView == Viewer2DView::Front ||
-          m_captureView == Viewer2DView::Side) &&
-         !highlight && !selected);
-    bool placedInstance = false;
-    if (useSymbolInstancing && m_captureCanvas && !skipCapture) {
-      std::string modelKey = NormalizeModelKey(gdtfPath);
-      if (modelKey.empty() && !f.gdtfSpec.empty())
-        modelKey = NormalizeModelKey(f.gdtfSpec);
-      if (modelKey.empty() && !f.typeName.empty())
-        modelKey = f.typeName;
-      if (modelKey.empty())
-        modelKey = "unknown";
-
-      if (!modelKey.empty()) {
-        SymbolKey symbolKey;
-        symbolKey.modelKey = modelKey;
-        symbolKey.viewKind = resolveSymbolView(m_captureView);
-        symbolKey.styleVersion = 1;
-
-        const auto &symbol =
-            m_bottomSymbolCache.GetOrCreate(symbolKey, [&](const SymbolKey &,
-                                                         uint32_t symbolId) {
-              SymbolDefinition definition{};
-              definition.symbolId = symbolId;
-              auto localCanvas =
-                  CreateRecordingCanvas(definition.localCommands, false);
-              CanvasTransform transform{};
-              localCanvas->BeginFrame();
-              localCanvas->SetTransform(transform);
-
-              ICanvas2D *prevCanvas = m_captureCanvas;
-              Viewer2DView prevView = m_captureView;
-              bool prevCaptureOnly = m_captureOnly;
-              bool prevIncludeGrid = m_captureIncludeGrid;
-              m_captureCanvas = localCanvas.get();
-              m_captureView = prevView;
-              m_captureOnly = true;
-              m_captureIncludeGrid = false;
-
-              if (itg != m_resourceSyncState.loadedGdtf.end()) {
-                size_t partIndex = 0;
-                for (const auto &obj : itg->second) {
-                  m_captureCanvas->SetSourceKey(fixtureCaptureKey + "_part" +
-                                                std::to_string(partIndex));
-                  auto applyCapture =
-                      [objTransform = obj.transform](
-                          const std::array<float, 3> &p) {
-                        return TransformPoint(objTransform, p);
-                      };
-                  float partR = r;
-                  float partG = g;
-                  float partB = b;
-                  if (!is2DViewer && obj.isLens) {
-                    partR = 1.0f;
-                    partG = 0.78f;
-                    partB = 0.35f;
-                  }
-                  DrawMeshWithOutline(obj.mesh, partR, partG, partB,
-                                      RENDER_SCALE, false, false, 0.0f, 0.0f,
-                                      0.0f, wireframe, mode, applyCapture,
-                                      false);
-                  ++partIndex;
-                }
-              } else {
-                m_captureCanvas->SetSourceKey(fixtureCaptureKey);
-                DrawCubeWithOutline(0.2f, r, g, b, false, false, 0.0f, 0.0f,
-                                    0.0f, wireframe, mode,
-                                    [](const std::array<float, 3> &p) {
-                                      return p;
-                                    });
-              }
-
-              localCanvas->EndFrame();
-              definition.bounds = ComputeSymbolBounds(definition.localCommands);
-
-              m_captureCanvas = prevCanvas;
-              m_captureView = prevView;
-              m_captureOnly = prevCaptureOnly;
-              m_captureIncludeGrid = prevIncludeGrid;
-              return definition;
-            });
-
-        Transform2D instanceTransform =
-            BuildInstanceTransform2D(fixtureTransform, m_captureView);
-        m_captureCanvas->PlaceSymbolInstance(symbol.symbolId,
-                                             instanceTransform);
-        placedInstance = true;
-      }
-    }
-    suppressCapture = placedInstance;
-
-    auto drawFixtureGeometry = [&]() {
-      if (itg != m_resourceSyncState.loadedGdtf.end()) {
-        size_t partIndex = 0;
-        for (const auto &obj : itg->second) {
-          glPushMatrix();
-          if (m_captureCanvas && !skipCapture) {
-            // Capture each GDTF geometry as its own source so fills do not hide
-            // outlines from sibling parts when exported to PDF.
-            m_captureCanvas->SetSourceKey(fixtureCaptureKey + "_part" +
-                                          std::to_string(partIndex));
-          }
-          float m2[16];
-          MatrixToArray(obj.transform, m2);
-          // GDTF geometry offsets are defined relative to the fixture
-          // in meters. Only the vertex coordinates need unit scaling.
-          ApplyTransform(m2, false);
-          auto applyCapture =
-              [fixtureTransform, objTransform = obj.transform](
-                  const std::array<float, 3> &p) {
-                auto local = TransformPoint(objTransform, p);
-                return TransformPoint(fixtureTransform, local);
-              };
-          float partR = r;
-          float partG = g;
-          float partB = b;
-          if (!is2DViewer && obj.isLens) {
-            partR = 1.0f;
-            partG = 0.78f;
-            partB = 0.35f;
-          }
-          const bool drawUnlit = !is2DViewer && obj.isLens;
-          DrawMeshWithOutline(obj.mesh, partR, partG, partB, RENDER_SCALE,
-                              highlight, selected, cx, cy, cz, wireframe,
-                              mode, applyCapture, drawUnlit);
-          glPopMatrix();
-          ++partIndex;
-        }
-      } else {
-        DrawCubeWithOutline(0.2f, r, g, b, highlight, selected, cx, cy, cz,
-                            wireframe, mode, applyFixtureCapture);
-      }
-    };
-
-    if (suppressCapture) {
-      ICanvas2D *prevCanvas = m_captureCanvas;
-      bool prevCaptureOnly = m_captureOnly;
-      m_captureCanvas = nullptr;
-      m_captureOnly = false;
-      drawFixtureGeometry();
-      m_captureCanvas = prevCanvas;
-      m_captureOnly = prevCaptureOnly;
-    } else {
-      drawFixtureGeometry();
-    }
-
-    glPopMatrix();
-
-    if (m_captureCanvas && !skipCapture)
-      m_captureCanvas->SetSourceKey("unknown");
-  }
-  if (forceFixturesOnTop && depthEnabled)
-    glEnable(GL_DEPTH_TEST);
-
-
+  OpaqueObjectPass::Render(*this, context, visibleSet, getLayerColor,
+                           resolveSymbolView);
+  OpaqueTrussPass::Render(*this, context, visibleSet, getLayerColor,
+                          resolveSymbolView);
+  OpaqueFixturePass::Render(*this, context, visibleSet, getTypeColor,
+                            getLayerColor, resolveSymbolView);
 }
 
 void Viewer3DController::RenderOverlayFrame(const RenderFrameContext &context,
